requirements.txt
openai==0.28.1
streamlit==1.28.1
python-dotenv==1.0.0
langchain==0.0.333
python-docx==1.1.0